requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
authlib==1.6.3
certifi==2025.8.3
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.3
click==8.2.1 ; sys_platform != 'emscripten'
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==45.0.7
cyclopts==3.23.1
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0 ; python_full_version < '4'
docutils==0.22
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.12.0
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.2
mcp==1.13.1
mdurl==0.1.2
more-itertools==10.8.0
openai==1.104.2
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
parse==1.20.2
pathable==0.4.4
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyperclip==1.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.3
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0 ; sys_platform != 'emscripten'
websockets==15.0.1
werkzeug==3.1.1
win32-setctime==1.2.0 ; sys_platform == 'win32'
